update - added the rest of the website
added a hero section with a number of logos (LogoLink component) and added a skill section where I have a coloum for each of the skill categories (SkillColumn component). The data for filling the logo and the skills is also added in data.json file, so i can dynamically add new links and new skills to the list. Also added a call to action section at the bottom with an image and more information and another set of logos.