update - styled the project caroussel
Started styling the project caroussel by making mock-ups of the websites, adding classNames to the different elements to style them better and adjusting the placement and order of elements. I made the background and the main contrast color a var so I can easitly change the colors in the future.